config/schema.py
import graphene
import graphql_jwt
import users.schema
import school.schema
import sys_management.schema
class Query(school.schema.Query, sys_management.schema.Query, users.schema.Query, graphene.ObjectType):
    pass

class Mutation(users.schema.Mutation, graphene.ObjectType):
    token_auth = graphql_jwt.ObtainJSONWebToken().Field()
    verify_token = graphql_jwt.Verify.Field()
    refresh_token = graphql_jwt.Refresh.Field()
schema = graphene.Schema(query=Query, mutation=Mutation)